Exception handling - delete

+++ src/main/java/com/darkra/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.darkra.course.services;
 
 import com.darkra.course.entities.User;
 import com.darkra.course.repositories.UserRepository;
+import com.darkra.course.services.exceptions.DatabaseException;
 import com.darkra.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+        try{
+            userRepository.deleteById(id);
+
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
